float problem... special case saved

+++ attacker.h
@@ -17,12 +17,12 @@ public:
     int qNum;
     int queryNum;
     bool updated;
-    Attacker(floatCounter * flowPara, FlowRuleTable *flowRuleTable, floatCounter * TTL, int mSize, int initialStateNum, double interval, double unit, double delta):model3(flowPara,flowRuleTable,TTL,mSize,initialStateNum,interval,unit,delta){
+    Attacker(floatCounter * flowPara, FlowRuleTable *flowRuleTable, floatCounter * TTL, int mSize, int initialStateNum, long double interval, long double unit, long double delta):model3(flowPara,flowRuleTable,TTL,mSize,initialStateNum,interval,unit,delta){
         //StateType i,j=1;
         updated=true;
     };
     void run(int qNum0,int flowInterest,StateType initialStateNum,std::set<std::set<int>>&attackFlow, MatD &PrXQ,VecD &IG);
-    double flowProbCompute3M(StateProb2& stateProb0,std::set<int> &queryInterest,int queryResult);
+    long double flowProbCompute3M(StateProb2& stateProb0,std::set<int> &queryInterest,int queryResult);
     void conditionalEntropyComputeM3(int flowInterest,int initialStateNum,VecD&conditionalEntropyQ,MatD&PrXQ);
     // void run(int qNum0,int flowInterest,StateType initialStateNum,std::set<int>&attackFlow, MatD&PrXQ,VecD&IG);
